requirements.txt
dash==0.26.4
dash-core-components==0.28.3
dash-html-components==0.12.0
dash-renderer==0.13.2
gunicorn==19.9.0  # for the deployment on Heroku
plotly==3.2.0
pytest==3.8.0
python-dotenv==0.9.1
PyInstaller==3.4
Flask-Assets==0.12
Flask-Uploads==0.2.1
